Restructured benchmark classes
