+++ src/jmh/java/ai/libs/sqlrest/benchmarks/IterativeSelect.java
@@ -1,15 +1,11 @@
-package ai.libs.sqlrest;
+package ai.libs.sqlrest.benchmarks;
 
 import ai.libs.jaicore.basic.kvstore.KVStore;
+import ai.libs.jaicore.db.sql.SQLAdapter;
+import ai.libs.sqlrest.*;
 import ai.libs.sqlrest.model.SQLQuery;
-import com.fasterxml.jackson.databind.ObjectMapper;
-import com.fasterxml.jackson.databind.type.CollectionType;
 import org.aeonbits.owner.ConfigCache;
-import org.apache.http.client.methods.CloseableHttpResponse;
-import org.apache.http.client.methods.HttpPost;
-import org.apache.http.entity.StringEntity;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
+import org.api4.java.datastructure.kvstore.IKVStore;
 import org.openjdk.jmh.annotations.*;
 import org.openjdk.jmh.infra.Blackhole;
 import org.slf4j.Logger;
@@ -23,14 +19,16 @@ import java.sql.SQLException;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
-@BenchmarkMode(Mode.AverageTime)
+/*
+ * Assumes that the service is running
+ */
+@BenchmarkMode(Mode.SingleShotTime)
 @OutputTimeUnit(TimeUnit.MILLISECONDS)
 @State(Scope.Benchmark)
 @Fork(value = 2, jvmArgs = {"-Xms4G", "-Xmx8G"})
-@Warmup(iterations = 1, time = 10)
-@Measurement(iterations = 4, time = 15)
-public class HttpClientPerformance extends AbstractServiceBenchmark {
-
+@Warmup(iterations = 6, time = 2)
+@Measurement(iterations = 10, time = 8)
+public class IterativeSelect  {
 
     private static final Logger logger = LoggerFactory.getLogger(IterativeSelect.class);
 
@@ -41,17 +39,22 @@ public class HttpClientPerformance extends AbstractServiceBenchmark {
     private static final IServerConfig SERVER_CONFIG = ConfigCache.getOrCreate(IServerConfig.class);
 
     @Param({
+            "1-random-time-null",
+            "100-random-time-null",
+            "1-random",
+            "100-random",
+            "1-time-null",
+            "100-time-null",
             "select-1",
             "select-100",
-            "select-10000",
+            "1-random-time-null-join",
+            "100-random-time-null-join",
+//            "1-random-time-null-subquery",
+//            "100-random-time-null-subquery"
     }) // each entry is 3.5 kByte
     private String query;
 
     private SQLQuery queryObj;
-
-    private CloseableHttpClient httpclient;
-    ObjectMapper mapper = new ObjectMapper();
-    CollectionType collectionType;
 
     private void createQuery() {
         String tableName = BENCHMARK_CONFIG.getBenchmarkTable();
@@ -60,14 +63,14 @@ public class HttpClientPerformance extends AbstractServiceBenchmark {
     }
 
     @Setup
-    public void setup() throws SQLException {
+    public void setup(SQLRestServiceState serviceState, SQLClientState state) {
         createQuery();
-        httpclient = HttpClients.createDefault();
-        collectionType = mapper.getTypeFactory().constructCollectionType(List.class, KVStore.class);
+        AbstractServiceBenchmark.startService(serviceState, state);
     }
 
+
     @Benchmark
-    public void springWebClient(SQLClientState state, Blackhole bh) throws IOException {
+    public void singleQueryOverService(SQLClientState state, Blackhole bh) throws IOException {
         WebClient webClient = state.getWebClient();
         WebClient.RequestHeadersSpec<?> request
                 = webClient.post()
@@ -79,21 +82,21 @@ public class HttpClientPerformance extends AbstractServiceBenchmark {
     }
 
     @Benchmark
-    public void apacheHttpClient(SQLClientState state, Blackhole bh) throws IOException {
-        HttpPost httpPost = new HttpPost(BENCHMARK_CONFIG.getServiceHost() + "/query");
-        String postBody = mapper.writeValueAsString(queryObj);
-        httpPost.setEntity(new  StringEntity(postBody));
-        httpPost.setHeader("Content-Type", "application/json");
-        CloseableHttpResponse response2 = httpclient.execute(httpPost);
+    public void singleQueryOverSingleAdapter(SQLClientState state, Blackhole bh) throws SQLException {
+        List<IKVStore> end = state.getAdapter().getResultsOfQuery(queryObj.getQuery());
+        assert end != null;
+        bh.consume(end);
+    }
 
-        try {
-            assert response2.getStatusLine().getStatusCode() == 200;
-            List<KVStore> o = mapper.readValue(response2.getEntity().getContent(), collectionType);
-            assert o != null;
-            bh.consume(o);
-        } finally {
-            response2.close();
-        }
+
+    @Benchmark
+    public void singleQueryOverFreshAdapters(SQLClientState state, Blackhole bh) throws SQLException {
+        SQLAdapter freshAdapter = new SQLAdapter(SERVER_CONFIG.getDBHost(),
+                state.getUserName(), state.getPassword(), state.getDbName(), SERVER_CONFIG.getDBPropUseSsl());
+        List<IKVStore> end = freshAdapter.getResultsOfQuery(queryObj.getQuery());
+        assert end != null;
+        bh.consume(end);
+        freshAdapter.close();
     }
 
 }
